added enemy health & HealthBar visualisation; towers now deal damage to enemies

+++ src/ProjectileStats.h
@@ -9,5 +9,5 @@ struct ProjectileStats
     }
 
     float move_speed;
-    float damage;
+    int damage;
 };
